Revert "Introducing Signature check in Masquerade (1/3)"
This is all simply absurd. The service is already protected in
the manifest with the MANAGE_OVERLAYS permission. It is reasonable
to conclude that if a user has granted said permission to an app
the user would also wish for the app to access resources to carry
out that function.

This reverts commit a872943c5bbf818ca090ae59715a6b93d747fa59.

Change-Id: Ia631ca0221bd37c4c4a81c8b29eeb066175aef87